devtools/client/memory/test/chrome/test_DominatorTree_02.html

<!DOCTYPE HTML>
<html>
<!--
Test that we show a link to load more children when some (but not all) are loaded.
-->
<head>
    <meta charset="utf-8">
    <title>Tree component test</title>
    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
    <!-- Give the container height so that the whole tree is rendered. -->
    <div id="container" style="height: 900px;"></div>

    <pre id="test">
        <script src="head.js" type="application/javascript"></script>
        <script type="application/javascript">
         "use strict";
         window.onload = async function() {
           try {
             const container = document.getElementById("container");

             const root = makeTestDominatorTreeNode({ moreChildrenAvailable: true }, [
               makeTestDominatorTreeNode({}),
             ]);
             ok(root.children);
             ok(root.moreChildrenAvailable);

             const expanded = new Set();
             expanded.add(root.nodeId);

             await renderComponent(DominatorTreeComponent(immutableUpdate(
                TEST_DOMINATOR_TREE_PROPS, {
                   dominatorTree: immutableUpdate(TEST_DOMINATOR_TREE_PROPS.dominatorTree, {
                     expanded,
                     root,
                   }),
                })), container);

             ok(container.querySelector(".more-children"),
                "Should get a link to load more children");
           } catch (e) {
             ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
           } finally {
             SimpleTest.finish();
           }
         };
        </script>
    </pre>
</body>
</html>
